fuzz: r285 harnesses #27+#28 — §6.2.1 read-symbol LUT differential + wide-carrier lossless decode
Depth round (fuzz mode) hardening the round-284 read_symbol
primary-lookup-table rewrite and word-load BitReader.

read_symbol_lut_diff (#27) is a differential oracle on the §6.2.1
fast path: PrefixCode::read_symbol (256-entry primary table keyed on
the 8 peeked wire-order bits, >8-bit continuation walk resuming at
length 9, near-EOF per-bit fallback, MIN_LOOKUP_USED amortization
gate) runs in lockstep against the crate's own pre-table per-bit row
walk, kept as the new #[doc(hidden)] behaviour-neutral
PrefixCode::read_symbol_reference oracle (delegates to the existing
private walk). After every symbol the harness asserts the decoded
symbol / typed refusal (including the PrefixError::Eof bit_pos /
wanted / available fields), the cursor bit position, and the alphabet
bound identical. The code under test is built off the wire
(PrefixCode::read at a fuzz-selected §6.2.3 alphabet, rest of the
stream as symbol soup) or from fuzz-shaped lengths repaired to an
exact §6.2.1 Kraft sum (greedy front fill + binary-decomposition tail
fill), so the mutator steers the used-symbol count across the
table-build gate and the length profile across the ≤8-bit fast path,
the >8-bit continuation rows, and the 15-bit ceiling. The §6.2.1
single-leaf tree is compared once (consumes no bits).

decode_lossless_lut (#28) re-drives decode_lossless /
decode_lossless_headerless at carrier dimensions widened into [1, 64]
(the round-273 sibling clamps at [1, 8], so its accepted streams read
only a handful of symbols per code), corpus-seeded from the fixture
corpus's VP8L chunk payloads plus entropy-heavy
reference-encoder-produced 64x64 noise / gradient / plasma tiles whose
§6.2 groups carry 100+-symbol codes with 9..15-bit tails — the LUT
fast path, continuation walk, and word-load
read_bits/peek_bits/advance_bits run hot inside the assembled pipeline
under adversarial mutation; the round-273 carrier-echo / pixel-count /
replay-determinism contract is asserted unchanged.

Campaigns (ASan, sequential): read_symbol_lut_diff 36,149,396 execs /
15 min, decode_lossless_lut 16,800,470 execs / 15 min, plus 4-minute
regression re-runs of prefix_code (32.2 M), decode_lossless (9.4 M),
and prefix_code_group (24.2 M) — zero divergences, zero crashes, no
findings: the round-284 fast path survived both oracles unchanged.
New unit test read_symbol_reference_matches_fast_path_on_lut_built_code
pins the differential on a 256-symbol code mixing ≤8-bit and >8-bit
lengths; full suite 435 lib tests + integration green.

Also: fuzz.yml stale "5 targets" comment refreshed (auto-discovery,
twenty-eight targets, ~64 s each); pre-existing clippy
doc_lazy_continuation / unnecessary_cast lints in the
inverse_subtract_green_indexing + parse_container harness docs fixed
so the fuzz sub-workspace is clippy -D warnings clean end to end.

+++ fuzz/fuzz_targets/inverse_subtract_green_indexing.rs
@@ -54,13 +54,14 @@
 //! via whichever residual buffer the upstream §5 entropy decoder produces
 //! and whichever color-table-delta + packed-index buffers the upstream
 //! sub-image decoders produce, all of which are bounded by the entropy
-//! stream itself. This sixteenth harness drives the §4.3 subtract-green
-//! + §4.4 inverse-color-table + §4.4 inverse-color-indexing transforms
-//! directly across their full attacker-reachable input cross-products
-//! within bounded sizes, with the §4.3 alpha/green preservation contract,
-//! the §4.4 color-table seed-preservation contract, and the §4.4
-//! color-indexing palette-lookup contract cross-checked against the
-//! RFC 9649 §4.3 + §4.4 spec text.
+//! stream itself. This sixteenth harness drives the §4.3
+//! subtract-green + §4.4 inverse-color-table + §4.4
+//! inverse-color-indexing transforms directly across their full
+//! attacker-reachable input cross-products within bounded sizes, with
+//! the §4.3 alpha/green preservation contract, the §4.4 color-table
+//! seed-preservation contract, and the §4.4 color-indexing
+//! palette-lookup contract cross-checked against the RFC 9649 §4.3 +
+//! §4.4 spec text.
 //!
 //! The contract under test, per RFC 9649 §4.3 + §4.4:
 //!
@@ -417,11 +418,7 @@ fuzz_target!(|data: &[u8]| {
     // is out of range. Cross-check by re-deriving the per-pixel index
     // from the packed buffer using the spec-text formula and comparing
     // to the impl's output.
-    let bits_per_index: u32 = if width_bits == 0 {
-        8
-    } else {
-        (8 / count) as u32
-    };
+    let bits_per_index: u32 = if width_bits == 0 { 8 } else { 8 / count };
     for y in 0..(height as usize) {
         for x in 0..(orig_width as usize) {
             let packed_x = x / (count as usize);
